+++ assets/src/Components/ImportModal.js
@@ -1,4 +1,5 @@
 /*global tiobDash*/
+/* eslint-disable no-console */
 import {
 	cleanupImport,
 	importContent,
@@ -12,11 +13,14 @@ import ImportStepper from './ImportStepper';
 import ImportModalNote from './ImportModalNote';
 import classnames from 'classnames';
 import ImportModalError from './ImportModalError';
+import ImportModalMock from './ImportModalMock';
+import CustomTooltip from './CustomTooltip';
 
-import { useState, useEffect, Fragment } from '@wordpress/element';
+import { useState, useEffect } from '@wordpress/element';
 import { withSelect, withDispatch } from '@wordpress/data';
 import { __, sprintf } from '@wordpress/i18n';
 import { compose } from '@wordpress/compose';
+
 import {
 	Button,
 	Icon,
@@ -32,6 +36,7 @@ const ImportModal = ( { setModal, editor, siteData, runTemplateImport } ) => {
 		content: true,
 		customizer: true,
 		widgets: true,
+		cleanup: false,
 	} );
 	const [ cleanupProgress, setCleanupProgress ] = useState( false );
 	const [ pluginsProgress, setPluginsProgress ] = useState( false );
@@ -49,7 +54,9 @@ const ImportModal = ( { setModal, editor, siteData, runTemplateImport } ) => {
 	const [ optionsOpened, setOptionsOpened ] = useState( true );
 
 	const { license, cleanupAllowed } = tiobDash;
-	const [ isCleanupAllowed, setIsCleanupAllowed ] = useState( cleanupAllowed );
+	const [ isCleanupAllowed, setIsCleanupAllowed ] = useState(
+		cleanupAllowed
+	);
 
 	useEffect( () => {
 		//const fetchAddress = siteData.remote_url || siteData.url;
@@ -81,10 +88,10 @@ const ImportModal = ( { setModal, editor, siteData, runTemplateImport } ) => {
 					const defaultOff =
 						result.default_off_recommended_plugins || [];
 
-					Object.keys( mandatory ).map( ( key ) => {
+					Object.keys( mandatory ).forEach( ( key ) => {
 						mandatory[ key ] = true;
 					} );
-					Object.keys( optional ).map( ( key ) => {
+					Object.keys( optional ).forEach( ( key ) => {
 						optional[ key ] = ! defaultOff.includes( key );
 					} );
 
@@ -96,7 +103,7 @@ const ImportModal = ( { setModal, editor, siteData, runTemplateImport } ) => {
 					setFetching( false );
 				} );
 			} )
-			.catch( ( error ) => {
+			.catch( () => {
 				setError( {
 					message: __(
 						'Something went wrong while loading the site data. Please refresh the page and try again.',
@@ -108,56 +115,6 @@ const ImportModal = ( { setModal, editor, siteData, runTemplateImport } ) => {
 			} );
 	}, [] );
 
-	const renderMock = () => {
-		return (
-			<Fragment>
-				<div className="modal-body">
-					<div className="header">
-						<span
-							className="title is-loading"
-							style={ { height: 35, marginBottom: 20 } }
-						/>
-						<p className="description is-loading" />
-						<p className="description is-loading" />
-					</div>
-					<div className="well is-loading">
-						<span
-							className="title is-loading"
-							style={ { height: 20 } }
-						/>
-						<ol>
-							<li />
-							<li />
-						</ol>
-					</div>
-
-					<div className="modal-toggles components-panel">
-						{ [ 1, 2 ].map( ( i ) => (
-							<div
-								key={ i }
-								className="components-panel__body options general is-opened"
-							>
-								<span className="title is-loading" />
-								<ul>
-									{ [ 1, 2, 3 ].map( ( i ) => (
-										<li className="option-row" key={ i }>
-											<div className="mock-icon is-loading" />
-											<span className="is-loading" />
-											<div className="toggle is-loading" />
-										</li>
-									) ) }
-								</ul>
-							</div>
-						) ) }
-					</div>
-				</div>
-				<div className="modal-footer">
-					<span className="link is-loading" />
-					<Button isPrimary className="import is-loading" />
-				</div>
-			</Fragment>
-		);
-	};
 	const Note = () => {
 		return (
 			<ImportModalNote
@@ -167,37 +124,26 @@ const ImportModal = ( { setModal, editor, siteData, runTemplateImport } ) => {
 		);
 	};
 
-	const ModalHead = () => {
-		if ( fetching ) {
-			return (
-				<>
-					<h1 className="is-loading title" />
-					<p className="is-loading description" />
-				</>
-			);
-		}
-		return (
-			<div className="header">
-				<h1>
-					{ sprintf(
-						/* translators: name of starter site */
-						__(
-							'Import %s as a complete site',
-							'templates-patterns-collection'
-						),
-						importData.title
-					) }
-				</h1>
-				<p className="description">
-					{ __(
-						'Import the entire site including customizer options, pages, content and plugins.',
+	const ModalHead = () => (
+		<div className="header">
+			<h1>
+				{ sprintf(
+					/* translators: name of starter site */
+					__(
+						'Import %s as a complete site',
 						'templates-patterns-collection'
-					) }
-				</p>
-			</div>
-		);
-	};
-
+					),
+					importData.title
+				) }
+			</h1>
+			<p className="description">
+				{ __(
+					'Import the entire site including customizer options, pages, content and plugins.',
+					'templates-patterns-collection'
+				) }
+			</p>
+		</div>
+	);
 	const Options = () => {
 		let map = {
 			content: {
@@ -217,10 +163,17 @@ const ImportModal = ( { setModal, editor, siteData, runTemplateImport } ) => {
 		if ( isCleanupAllowed === 'yes' ) {
 			map = {
 				cleanup: {
-					title: __( 'Cleanup previous Import', 'templates-patterns-collection' ),
-					icon: 'image-rotate',
+					icon: 'trash',
+					title: __(
+						'Cleanup previous import',
+						'templates-patterns-collection'
+					),
+					tooltip: __(
+						'This will remove any plugins, images, customizer options, widgets posts and pages added by the previous demo import',
+						'templates-patterns-collection'
+					),
 				},
-				...map
+				...map,
 			};
 		}
 
@@ -239,15 +192,18 @@ const ImportModal = ( { setModal, editor, siteData, runTemplateImport } ) => {
 				) }
 			>
 				{ Object.keys( map ).map( ( id, index ) => {
+					const rowClass = classnames( 'option-row', {
+						active: general[ id ],
+					} );
+					const { icon, title, tooltip } = map[ id ];
+
 					return (
-						<PanelRow className="option-row" key={ index }>
-							<Icon
-								className={ classnames( {
-									active: general[ id ],
-								} ) }
-								icon={ map[ id ].icon }
-							/>
-							<span>{ map[ id ].title }</span>
+						<PanelRow className={ rowClass } key={ index }>
+							<Icon icon={ icon } />
+							<span>{ title }</span>
+							{ tooltip && (
+								<CustomTooltip>{ tooltip }</CustomTooltip>
+							) }
 							<div className="toggle-wrapper">
 								<ToggleControl
 									checked={ general[ id ] }
@@ -286,14 +242,12 @@ const ImportModal = ( { setModal, editor, siteData, runTemplateImport } ) => {
 				title={ __( 'Plugins', 'templates-patterns-collection' ) }
 			>
 				{ Object.keys( allPlugins ).map( ( slug, index ) => {
+					const rowClass = classnames( 'option-row', {
+						active: pluginOptions[ slug ],
+					} );
 					return (
-						<PanelRow className="option-row" key={ index }>
-							<Icon
-								icon="admin-plugins"
-								className={ classnames( {
-									active: pluginOptions[ slug ],
-								} ) }
-							/>
+						<PanelRow className={ rowClass } key={ index }>
+							<Icon icon="admin-plugins" />
 							<span
 								dangerouslySetInnerHTML={ {
 									__html: allPlugins[ slug ],
@@ -323,10 +277,9 @@ const ImportModal = ( { setModal, editor, siteData, runTemplateImport } ) => {
 
 	function runImportCleanup() {
 		console.clear();
-		console.log( general );
 		if ( ! general.cleanup ) {
 			console.log( '[S] Cleanup.' );
-			runImport()
+			runImport();
 			return false;
 		}
 		setCurrentStep( 'cleanup' );
@@ -341,7 +294,9 @@ const ImportModal = ( { setModal, editor, siteData, runTemplateImport } ) => {
 				setCleanupProgress( 'done' );
 				runImport();
 			} )
-			.catch( ( error ) => handleError( error, 'cleanup' ) );
+			.catch( ( incomingError ) =>
+				handleError( incomingError, 'cleanup' )
+			);
 	}
 
 	function runImport() {
@@ -363,7 +318,9 @@ const ImportModal = ( { setModal, editor, siteData, runTemplateImport } ) => {
 				setPluginsProgress( 'done' );
 				runImportContent();
 			} )
-			.catch( ( error ) => handleError( error, 'plugins' ) );
+			.catch( ( incomingError ) =>
+				handleError( incomingError, 'plugins' )
+			);
 	}
 
 	function runImportContent() {
@@ -394,7 +351,9 @@ const ImportModal = ( { setModal, editor, siteData, runTemplateImport } ) => {
 				setContentProgress( 'done' );
 				runImportCustomizer();
 			} )
-			.catch( ( error ) => handleError( error, 'content' ) );
+			.catch( ( incomingError ) =>
+				handleError( incomingError, 'content' )
+			);
 	}
 
 	function runImportCustomizer() {
@@ -419,7 +378,9 @@ const ImportModal = ( { setModal, editor, siteData, runTemplateImport } ) => {
 				setCustomizerProgress( 'done' );
 				runImportWidgets();
 			} )
-			.catch( ( error ) => handleError( error, 'customizer' ) );
+			.catch( ( incomingError ) =>
+				handleError( incomingError, 'customizer' )
+			);
 	}
 
 	function runImportWidgets() {
@@ -439,7 +400,9 @@ const ImportModal = ( { setModal, editor, siteData, runTemplateImport } ) => {
 				setWidgetsProgress( 'done' );
 				importDone();
 			} )
-			.catch( ( error ) => handleError( error, 'widgets' ) );
+			.catch( ( incomingError ) =>
+				handleError( incomingError, 'widgets' )
+			);
 	}
 
 	function importDone() {
@@ -449,7 +412,7 @@ const ImportModal = ( { setModal, editor, siteData, runTemplateImport } ) => {
 		setImporting( false );
 	}
 
-	function handleError( error, step ) {
+	function handleError( incomingError, step ) {
 		setImporting( false );
 		setCurrentStep( null );
 		if ( 'cleanup' === step ) {
@@ -506,10 +469,10 @@ const ImportModal = ( { setModal, editor, siteData, runTemplateImport } ) => {
 				break;
 		}
 		setError(
-			error.data
+			incomingError.data
 				? {
 						message: map[ step ],
-						code: error.data,
+					code: incomingError.data,
 				  }
 				: { message: map[ step ] }
 		);
@@ -521,8 +484,6 @@ const ImportModal = ( { setModal, editor, siteData, runTemplateImport } ) => {
 		}
 		setModal( false );
 	};
-
-	const runTemplatesImport = () => {};
 
 	const externalPluginsInstalled = siteData.external_plugins
 		? siteData.external_plugins.every( ( value ) => true === value.active )
@@ -548,9 +509,9 @@ const ImportModal = ( { setModal, editor, siteData, runTemplateImport } ) => {
 			isDismissible={ ! importing && ! fetching }
 		>
 			{ fetching ? (
-				renderMock()
+				<ImportModalMock />
 			) : (
-				<Fragment>
+				<>
 					<div className="modal-body">
 						{ ! importing && 'done' !== currentStep && ! error ? (
 							<>
@@ -586,7 +547,7 @@ const ImportModal = ( { setModal, editor, siteData, runTemplateImport } ) => {
 									/>
 								) }
 								{ 'done' === currentStep && (
-									<Fragment>
+									<>
 										<hr />
 										<p className="import-result">
 											{ __(
@@ -595,7 +556,7 @@ const ImportModal = ( { setModal, editor, siteData, runTemplateImport } ) => {
 											) }
 										</p>
 										<hr />
-									</Fragment>
+									</>
 								) }
 							</>
 						) }
@@ -603,7 +564,7 @@ const ImportModal = ( { setModal, editor, siteData, runTemplateImport } ) => {
 					{ ! importing && (
 						<div className="modal-footer">
 							{ 'done' !== currentStep ? (
-								<Fragment>
+								<>
 									<Button
 										className="import-templates"
 										isLink
@@ -633,7 +594,7 @@ const ImportModal = ( { setModal, editor, siteData, runTemplateImport } ) => {
 											) }
 										</Button>
 									) }
-								</Fragment>
+								</>
 							) : (
 								<div className="import-done-actions">
 									<Button
@@ -669,7 +630,7 @@ const ImportModal = ( { setModal, editor, siteData, runTemplateImport } ) => {
 							) }
 						</div>
 					) }
-				</Fragment>
+				</>
 			) }
 		</Modal>
 	);